+++ src/app/sw.ts
@@ -1,0 +1,20 @@
+import { defaultCache } from "@serwist/next/worker";
+import type { PrecacheEntry } from "@serwist/precaching";
+import { installSerwist } from "@serwist/sw";
+
+declare global {
+  interface WorkerGlobalScope {
+    // Change this attribute's name to your `injectionPoint`
+    __SW_MANIFEST: (PrecacheEntry | string)[] | undefined;
+  }
+}
+
+declare const self: WorkerGlobalScope;
+
+installSerwist({
+  precacheEntries: self.__SW_MANIFEST,
+  skipWaiting: true,
+  clientsClaim: true,
+  navigationPreload: true,
+  runtimeCaching: defaultCache,
+});